Initial Test Framework Commit
The object and associated method found in Test.py can be used for
rapidly testing a single method. The file testTest.py was used for some
initial (and rather basic) testing of the Test.py file. More changes
are likely to be necessary, as this does not currently account for
methods that return none and modify state (i.e object setters).