Initial Help Menu Commit
The basic help menu is done, this could serve as an outline for the
command structure used in the controller, and will certainly inform
the way the interpreter is built. The issues that remain are as follows:
1) when loading, where do we search? Do we require an absolute filepath,
do we just search the current directory, do we have support for both?
2) do we want to support all the list functions provided in the help menu? I added a few extra because they seem easy enough, but time may be a factor.
3) is the format of the messages in the help menu acceptable?